pyproject.toml
[project]
name = "langgraph-rag-agent-chat"
version = "0.1.0"
description = "Production-ready RAG agent with LangGraph & LangChain. Document upload, vector search, and intelligent chat interface."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.1",
    "gunicorn>=23.0.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "pydantic-settings>=2.11.0",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "psycopg2-binary>=2.9.9",
    "pymilvus>=2.4.0",
    "pgvector>=0.4.1",
    "langgraph-checkpoint-postgres>=3.0.1",
    "asgiref>=3.11.0",
]
